Refactoring and update linearizability tests (#353)
1) Extract common utility functions from SkvbcChaosTest to SkvbcTracker.
2) Move Class Status from "test_skvbc_linearizability.py" to "skvbc_history_tracker" - status class helps us to track reads and writes so its main usage is from history tracker.
