+++ tests/util/skvbc_history_tracker.py
@@ -12,6 +12,9 @@
 
 import time
 from enum import Enum
+from util import skvbc as kvbc
+import trio
+import random
 
 from util.skvbc_exceptions import(
     ConflictingBlockWriteError,
@@ -20,6 +23,8 @@ from util.skvbc_exceptions import(
     InvalidReadError,
     PhantomBlockError
 )
+
+MAX_LOOKBACK=10
 
 class SkvbcWriteRequest:
     """
@@ -196,6 +201,37 @@ class Block:
            f'  kvpairs={self.kvpairs}\n'
            f'  req_index={self.req_index}\n')
 
+class Status:
+    """
+    Status about the order of writes and reads.
+    This is useful for debugging linearizability check failures.
+    """
+    def __init__(self, config):
+        self.config = config
+        self.start_time = time.monotonic()
+        self.end_time = 0
+        self.last_client_reply = 0
+        self.client_timeouts = {}
+        self.client_replies = {}
+
+    def record_client_reply(self, client_id):
+        self.last_client_reply = time.monotonic()
+        count = self.client_replies.get(client_id, 0)
+        self.client_replies[client_id] = count + 1
+
+    def record_client_timeout(self, client_id):
+        count = self.client_timeouts.get(client_id, 0)
+        self.client_timeouts[client_id] = count + 1
+
+    def __str__(self):
+        return (f'{self.__class__.__name__}:\n'
+           f'  config={self.config}\n'
+           f'  test_duration={self.end_time - self.start_time} seconds\n'
+           f'  time_since_last_client_reply='
+           f'{self.end_time - self.last_client_reply} seconds\n'
+           f'  client_timeouts={self.client_timeouts}\n'
+           f'  client_replies={self.client_replies}\n')
+
 class SkvbcTracker:
     """
     Track requests, expected and actual responses from SimpleKVBC test
@@ -314,7 +350,7 @@ class SkvbcTracker:
     clusters with lots of blocks, but we may want to add it as an optional check
     in the future.
     """
-    def __init__(self, initial_kvpairs={}):
+    def __init__(self, initial_kvpairs={}, skvbc=None, bft_network=None):
         # A partial order of all requests (SkvbcWriteRequest | SkvbcReadRequest)
         # issued against SimpleKVBC.  History tracks requests and responses. A
         # happens-before relationship exists between responses and requests
@@ -350,6 +386,13 @@ class SkvbcTracker:
         # Blocks that get filled in by the call to fill_missing_blocks
         # These blocks were created by write requests that never got replies.
         self.filled_blocks = {}
+
+        self.skvbc = skvbc
+
+        self.bft_network = bft_network
+
+        if self.bft_network is not None:
+            self.status = Status(bft_network.config)
 
     def send_write(self, client_id, seq_num, readset, writeset, read_block_id):
         """Track the send of a write request"""
@@ -724,3 +767,91 @@ class SkvbcTracker:
         causal_state = self.outstanding[(rpy.client_id, rpy.seq_num)]
         index = causal_state.req_index
         return (self.history[index], index)
+
+    async def fill_missing_blocks_and_verify(self):
+         try:
+             # Use a new client, since other clients may not be responsive due to
+             # past failed responses.
+             client = await self.skvbc.bft_network.new_client()
+             last_block_id = await self.get_last_block_id(client)
+             print(f'Last Block ID = {last_block_id}')
+             missing_block_ids = self.get_missing_blocks(last_block_id)
+             print(f'Missing Block IDs = {missing_block_ids}')
+             blocks = await self.get_blocks(client, missing_block_ids)
+             self.fill_missing_blocks(blocks)
+             self.verify()
+         except Exception as e:
+             print(f'retries = {client.retries}')
+             self.status.end_time = time.monotonic()
+             print("HISTORY...")
+             for i, entry in enumerate(self.history):
+                 print(f'Index = {i}: {entry}\n')
+             print("BLOCKS...")
+             print(f'{self.blocks}\n')
+             print(str(self.status), flush=True)
+             print("FAILURE...")
+             raise(e)
+
+    async def get_blocks(self, client, block_ids):
+        blocks = {}
+        for block_id in block_ids:
+            retries = 12 # 60 seconds
+            for i in range(0, retries):
+                try:
+                    msg = kvbc.SimpleKVBCProtocol.get_block_data_req(block_id)
+                    blocks[block_id] = kvbc.SimpleKVBCProtocol.parse_reply(await client.read(msg))
+                    break
+                except trio.TooSlowError:
+                    if i == retries - 1:
+                        raise
+            print(f'Retrieved block {block_id}')
+        return blocks
+
+    async def get_last_block_id(self, client):
+        msg = kvbc.SimpleKVBCProtocol.get_last_block_req()
+        return kvbc.SimpleKVBCProtocol.parse_reply(await client.read(msg))
+
+    async def send_tracked_write(self, client, max_set_size):
+        readset = self.readset(0, max_set_size)
+        writeset = self.writeset(max_set_size)
+        read_version = self.read_block_id()
+        msg = self.skvbc.write_req(readset, writeset, read_version)
+        seq_num = client.req_seq_num.next()
+        client_id = client.client_id
+        self.send_write(
+            client_id, seq_num, readset, dict(writeset), read_version)
+        try:
+            serialized_reply = await client.write(msg, seq_num)
+            self.status.record_client_reply(client_id)
+            reply = self.skvbc.parse_reply(serialized_reply)
+            self.handle_write_reply(client_id, seq_num, reply)
+        except trio.TooSlowError:
+            self.status.record_client_timeout(client_id)
+            return
+
+    async def send_tracked_read(self, client, max_set_size):
+        readset = self.readset(1, max_set_size)
+        msg = self.skvbc.read_req(readset)
+        seq_num = client.req_seq_num.next()
+        client_id = client.client_id
+        self.send_read(client_id, seq_num, readset)
+        try:
+            serialized_reply = await client.read(msg, seq_num)
+            self.status.record_client_reply(client_id)
+            reply = self.skvbc.parse_reply(serialized_reply)
+            self.handle_read_reply(client_id, seq_num, reply)
+        except trio.TooSlowError:
+            self.status.record_client_timeout(client_id)
+            return
+
+    def read_block_id(self):
+        start = max(0, self.last_known_block - MAX_LOOKBACK)
+        return random.randint(start, self.last_known_block)
+
+    def readset(self, min_size, max_size):
+        return self.skvbc.random_keys(random.randint(min_size, max_size))
+
+    def writeset(self, max_size):
+        writeset_keys = self.skvbc.random_keys(random.randint(0, max_size))
+        writeset_values = self.skvbc.random_values(len(writeset_keys))
+        return list(zip(writeset_keys, writeset_values))
